experiments/xiaoke_doc_assist/requirements.txt
streamlit
openai
python-dotenv
PyMuPDF
